pyproject.toml
[tool.poetry]
name = "tmatch"
version = "0.1.0"
description = ""
authors = ["Hamza Farhan <thehamza96@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
notebook = "^7.1.0"
ipython = "^8.22.1"
pytest = "^8.0.1"
ruff = "^0.2.2"
langchain = "^0.1.9"
termcolor = "^2.4.0"
tiktoken = "^0.6.0"
litellm = "^1.27.1"
redis = "^5.0.1"
google-generativeai = "^0.3.2"
google-cloud-aiplatform = "^1.42.1"
pandas = "^2.2.1"
python-docx = "^1.1.0"
sentence-transformers = "^2.4.0"
demoji = "^1.1.0"
unstructured = {extras = ["pdf"], version = "^0.12.4"}
fastcore = "^1.5.29"
ipywidgets = "^8.1.2"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.27.1"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
